modern

+++ database/migrations/2024_08_25_053031_update_rooms_table.php
@@ -30,16 +30,33 @@ return new class extends Migration
      */
     public function down(): void
     {
-        //
         Schema::table('rooms', function (Blueprint $table) {
-            $table->dropForeign(['building_id']);
-            $table->dropColumn('building_id');
-            $table->dropColumn('category_PB');
-            $table->dropColumn('RadiationSafetyZone');
-            $table->dropForeign(['owner_division']);
-            $table->dropColumn('owner_division');
-            $table->dropForeign(['owner_subdivision']);
-            $table->dropColumn('owner_subdivision');
+            if (Schema::hasColumn('rooms', 'building_id')) {
+                // Замініть 'rooms_building_id_foreign' на правильне ім'я зовнішнього ключа, якщо потрібно
+                $table->dropForeign(['building_id']);
+                $table->dropColumn('building_id');
+            }
+    
+            if (Schema::hasColumn('rooms', 'category_PB')) {
+                $table->dropColumn('category_PB');
+            }
+    
+            if (Schema::hasColumn('rooms', 'RadiationSafetyZone')) {
+                $table->dropColumn('RadiationSafetyZone');
+            }
+    
+            if (Schema::hasColumn('rooms', 'owner_division')) {
+                // Замініть 'rooms_owner_division_foreign' на правильне ім'я зовнішнього ключа, якщо потрібно
+                $table->dropForeign(['owner_division']);
+                $table->dropColumn('owner_division');
+            }
+    
+            if (Schema::hasColumn('rooms', 'owner_subdivision')) {
+                // Замініть 'rooms_owner_subdivision_foreign' на правильне ім'я зовнішнього ключа, якщо потрібно
+                $table->dropForeign(['owner_subdivision']);
+                $table->dropColumn('owner_subdivision');
+            }
         });
     }
+    
 };
